% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{get_data_1}
\alias{get_data_1}
\title{Get data 1}
\usage{
get_data_1(diffs = TRUE)
}
\arguments{
\item{diffs}{return the differentiated series or not? (lag = 1)}
}
\value{
a list of time series objects
}
\description{
Data from Task Views + synthetic
}
