% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/martingaletests.R
\name{martingale_test}
\alias{martingale_test}
\title{Martingale Test}
\usage{
martingale_test(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A list of p-values for the martingale tests.
}
\description{
Martingale Test
}
