% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetize.R
\name{rgaussiandens}
\alias{rgaussiandens}
\title{Simulate Gaussian kernel density}
\usage{
rgaussiandens(
  x,
  n = length(x),
  p = 1,
  seed = 123,
  method = c("antithetic", "traditional")
)
}
\description{
Simulate Gaussian kernel density
}
