% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsg2plus.R
\name{simulate_bsg2plus}
\alias{simulate_bsg2plus}
\title{Simulate Black-Scholes-G2++ model}
\usage{
simulate_bsg2plus(n = 100L, t0 = 1L, horizon = 5L, K = 95)
}
\arguments{
\item{n}{number of simulations}

\item{t0}{valuation date (must be an integer > 1)}

\item{T}{maturity}
}
\value{
a list with the simulated short rate,
asset price and discount bond price
}
\description{
Simulate Black-Scholes-G2++ model
}
\examples{

require(esgtoolkit)

sims <- simulatetimeseries::simulate_bsg2plus(t0=1, n=10L,
horizon = 5L)

par(mfrow=c(2,2))
esgtoolkit::esgplotbands(sims$discount_bond_prices, main="Discount bond prices", xlab="Maturity", ylab="Prices")
esgtoolkit::esgplotbands(window(sims$discount_rates, start=1), main="Discount rates", xlab="Maturity", ylab="Rates")
esgtoolkit::esgplotbands(sims$asset_price, main="Asset prices", xlab="Horizon", ylab="Prices")
esgtoolkit::esgplotbands(sims$short_rate, main="Short rates", xlab="Horizon", ylab="Rates")

}
