% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatecorrgaussian.R
\name{simulate_correlated_gaussian}
\alias{simulate_correlated_gaussian}
\title{Title}
\usage{
simulate_correlated_gaussian(n = 100L, cor_matrix = diag(3))
}
\arguments{
\item{cor_matrix}{}
}
\description{
Title
}
\examples{

cor_matrix <- matrix(c(1, 0.6, 0.6, 1), nrow = 2, byrow = TRUE)
correlated_gaussian <- simulate_correlated_gaussian(n=100,
cor_matrix=cor_matrix)
print(cor(correlated_gaussian))

}
