% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{simulate_time_series_3}
\alias{simulate_time_series_3}
\title{Simulate a univariate time series dataset 3}
\usage{
simulate_time_series_3(n = 100, seed = 123)
}
\arguments{
\item{n}{numerical, number of data points}

\item{seed}{int, reproducibility seed}
}
\value{
a native time series object
}
\description{
Simulate a univariate time series dataset 3
}
\examples{

print(simulate_time_series_3(10))

}
