% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{simulate_time_series_4}
\alias{simulate_time_series_4}
\title{Simulate a univariate time series dataset 4}
\usage{
simulate_time_series_4(n = 100, psi = 0.1, xi = 0.1, seed = 123)
}
\arguments{
\item{n}{numerical, number of data points}

\item{psi}{1st parameter for innovation variance ()}

\item{xi}{2nd parameter for innovation variance}

\item{seed}{int, reproducibility seed}
}
\value{
a native time series object
}
\description{
Simulate a univariate time series dataset 4
}
\examples{

plot(simulate_time_series_4())

}
