\name{simulatetimeseries-package}
\alias{simulatetimeseries-package}
\alias{simulatetimeseries}
\docType{package}
\title{\packageTitle{simulatetimeseries}}
\description{\packageDescription{simulatetimeseries}}
\section{Package Content}{\packageIndices{simulatetimeseries}}
\author{\packageAuthor{simulatetimeseries}}
\section{Maintainer}{\packageMaintainer{simulatetimeseries}}
\keyword{package}
